Fixed identity mapper test
This thest also failes to the slightly changed interpolation
results of itk v5 which result in not an exact dublicate image.
Due to the fact that the test should not check the interpolation,
bit if the mapR tool behaves correctly if no registration is
provided (generation identity transform), I have switched the
interpolation style to get rid of the problem for the test.